refactor: move CLI argument models out of validation into cli/args.py
Six validation/*_validator.py files held no validators — each was a
single pydantic CLI-argument model consumed only by cli/ commands.
Consolidate them into cli/args.py, update the importers, and drop the
BatchModeCompatibilityValidator alias, which had no users.
