+++ .changes/unreleased/Under the Hood-20260807-213000.yaml
@@ -1,0 +1,3 @@
+kind: Under the Hood
+body: "Move the six CLI argument models (RunCommandArgs, RetryCommandArgs, InitCommandArgs, CleanCommandArgs, StatusCommandArgs, BatchCommandArgs) from misnamed validation/*_validator.py files into one cli/args.py, and drop the unused BatchModeCompatibilityValidator alias. No behavior change."
+time: 2026-08-07T21:30:00Z
