+++ agent_actions/cli/clean.py
@@ -4,9 +4,9 @@ from pathlib import Path
 
 import click
 
+from agent_actions.cli.args import CleanCommandArgs
 from agent_actions.cli.cli_decorators import handles_user_errors, requires_project
 from agent_actions.llm.realtime.cleaner import Cleaner
-from agent_actions.validation.clean_validator import CleanCommandArgs
 
 
 @click.command(
